feat(alert-system): use parent_guid instead of correlation_id
- Remove correlation_id from alert notification flow
- Use parent_guid for alert notification tracking
- Disable parallel processing to keep execution deterministic
- Move email processing logic out of tasks.py
