+++ alert_system/management/commands/alert_notification.py
@@ -1,9 +1,14 @@
 from django.core.management.base import BaseCommand
 from sentry_sdk import monitor
 
+from alert_system.email_processing import process_email_alert
 from alert_system.models import LoadItem
-from alert_system.tasks import process_email_alert
 from main.sentry import SentryMonitor
+
+# NOTE: Disabled parallel processing to avoid inconsistent state and keep
+#  execution deterministic. Email logic is intentionally moved out of tasks.py for now.
+#  If reintroduced later, a Celery chain may be used to ensure proper ordering
+#  and retry management.
 
 
 class Command(BaseCommand):
@@ -18,9 +23,8 @@ class Command(BaseCommand):
             self.stdout.write(self.style.NOTICE("No eligible items found"))
             return
 
-        self.stdout.write(self.style.NOTICE(f"Queueing {items.count()} items for alert email notification."))
+        self.stdout.write(self.style.NOTICE(f"Processing {items.count()} items for alert email notification."))
 
         for item in items.iterator():
-            process_email_alert.delay(load_item_id=item.id)
-
-        self.stdout.write(self.style.SUCCESS("All alert notification email queued successfully"))
+            process_email_alert(load_item_id=item.id)
+        self.stdout.write(self.style.SUCCESS("All alert notification emails processed successfully"))
