+++ alert_system/models.py
@@ -280,8 +280,7 @@ class AlertEmailThread(models.Model):
         related_name="alert_email_threads",
     )
 
-    correlation_id = models.CharField(
-        max_length=255,
+    parent_guid = models.CharField(
         help_text=_("Identifier linking related LoadItems into the same email thread."),
     )
 
@@ -304,12 +303,13 @@ class AlertEmailThread(models.Model):
         verbose_name = _("Email Thread")
         verbose_name_plural = _("Email Threads")
         ordering = ["-id"]
+        constraints = [models.UniqueConstraint(fields=["parent_guid", "user"], name="unique_user_guid")]
         indexes = [
-            models.Index(fields=["correlation_id", "user"]),
+            models.Index(fields=["parent_guid", "user"]),
         ]
 
     def __str__(self):
-        return f"Thread: {self.user.get_full_name()}-{self.correlation_id}"
+        return f"Thread: {self.user.get_full_name()}-{self.parent_guid}"
 
 
 class AlertEmailLog(models.Model):
